requirements.txt
mysql-connector-python
customtkinter
bcrypt
reportlab
pytest
